# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Vedetta NVR API
  version: "0.1.0"
  description: API for managing cameras, events, recordings, and object detection in the Vedetta Network Video Recorder.
servers:
  - url: /
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
  /api/auth/login:
    post:
      operationId: login
      summary: Authenticate with username and password
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/auth/change-password:
    post:
      operationId: changePassword
      summary: Change the current user's password
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOk"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/tokens:
    get:
      operationId: listTokens
      summary: List API tokens for the authenticated user
      tags: [auth]
      responses:
        "200":
          description: Token list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenListResponse"
    post:
      operationId: createToken
      summary: Create a new API token
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTokenRequest"
      responses:
        "201":
          description: Token created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTokenResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        remember:
          type: boolean
    LoginResponse:
      type: object
      required: [username, kind, csrf_token, expires_at]
      properties:
        username:
          type: string
        kind:
          type: string
        csrf_token:
          type: string
        expires_at:
          type: string
          format: date-time
...
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
        new_password:
          type: string
    CreateTokenRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        scopes:
          type: array
          items:
            type: string
    CreateTokenResponse:
      type: object
      required: [id, name, token, token_prefix, scopes, created_at]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        token:
          type: string
        token_prefix:
          type: string
        scopes:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
...
    TokenListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TokenListItem"
        total:
          type: integer
    TokenListItem:
      type: object
      required: [id, name, token_prefix, scopes, created_at]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        token_prefix:
          type: string
        scopes:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    StatusOk:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
...